Update home page component imports
