+++ components/common/footer.tsx
