+++ lib/product/index.ts
@@ -1,0 +1,3 @@
+export * from "./product";
+export * from "./pharmacyProduct";
+export * from "./category";
