+++ lib/product/pharmacyProduct.ts
